revamp: search bar refactored
