--- src/components/atoms/SearchBar/SearchBar.module.css
@@ -1,108 +1,0 @@
-.searchBar {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-}
-
-.searchBar:focus-within {
-  outline: solid 0.125rem var(--nexus-blue);
-  border-radius: 2pt;
-  color: var(--midnight-blue);
-}
-
-.searchBar input[type='search'] {
-  height: 2.125rem;
-  background: var(--theme-search);
-  color: var(--midnight-blue);
-  border: none;
-  border-radius: 0.125rem 0 0 0.125rem;
-  box-shadow: var(--card-shadow-subtle);
-  box-sizing: border-box;
-  outline: none;
-  font-size: var(--font-normal);
-  width: 100%;
-  padding-left: 0.5rem;
-}
-
-.searchBar input[type='search']::-webkit-search-cancel-button {
-  position: relative;
-  width: 1.5rem;
-  height: 1.5rem;
-  background: url('../../../assets/icons/cross.svg');
-  background-repeat: no-repeat;
-  background-position: center;
-  -webkit-appearance: none;
-  cursor: pointer;
-}
-
-/* * This line cause placeholder to be invisible in chrome  */
-/* .searchBar input[type='search']::-webkit-input-placeholder, */
-.searchBar input[type='search']:-ms-input-placeholder,
-.searchBar input[type='search']::placeholder {
-  color: var(--theme-search-text);
-}
-
-.searchBar button {
-  height: 2.125rem;
-  background: var(--theme-search);
-  box-shadow: var(--card-shadow-subtle);
-  border-radius: 0 0.25rem 0.25rem 0;
-  border: none;
-  /* position: relative; */
-  right: 5%;
-  margin: 0;
-  padding: 0;
-  /* padding-top: 0.3rem; */
-  cursor: pointer;
-}
-.searchBar button:hover {
-  opacity: 0.95;
-}
-
-.searchBar button:active {
-  transform: scale(1.05);
-}
-
-.searchBar button .searchIcon {
-  width: 1.5rem;
-  height: 1.5rem;
-}
-
-.searchBarLong {
-  box-shadow: var(--card-shadow-subtle);
-  max-height: 3.25rem;
-}
-.searchBarLong:focus-within {
-  outline: none;
-}
-
-.searchBarLong input[type='search'],
-.searchBarLong .searchLong {
-  height: 2.625rem;
-  color: var(--midnight-blue);
-  border-radius: 0.125rem;
-  font-size: var(--font-normal);
-  font-family: inherit;
-  font-weight: normal;
-  width: 100%;
-  box-shadow: var(--card-shadow-subtle);
-}
-
-.searchBarLong input[type='search']:focus {
-  outline: solid 0.125rem var(--nexus-blue);
-  border-radius: 2pt;
-  color: var(--midnight-blue);
-}
-
-.searchBarLong .searchLongBtn {
-  height: 2.625rem;
-  margin-left: 0.25rem;
-  border-radius: 0.125rem;
-  width: 2.625rem;
-  box-shadow: var(--card-shadow-subtle);
-}
-.searchBarLong .searchLongBtn:focus {
-  outline: solid 0.125rem var(--nexus-blue);
-  border-radius: 2pt;
-}
